+++ src/components/Pagination/Pagination.tsx
@@ -1,0 +1,7 @@
+import React from 'react';
+
+const Pagination = () => {
+  return <div>Pagination</div>;
+};
+
+export default Pagination;
