+++ src/components/Pagination/index.tsx
@@ -1,0 +1,1 @@
+export { default as Pagination } from './Pagination';
