--- src/pages/api/post/[id].ts
@@ -1,9 +1,0 @@
-import type { NextApiRequest, NextApiResponse } from 'next';
-
-type Data = {
-  name: string;
-};
-
-export default function handler(req: NextApiRequest, res: NextApiResponse<Data>) {
-  res.status(200).json({ name: 'post detail' });
-}
