+++ tsconfig.json
@@ -25,7 +25,8 @@
       "@styles/*": ["src/styles/*"],
       "@type/*": ["src/type/*"],
       "@utils/*": ["src/utils/*"],
-      "@recoil/*": ["src/recoil/*"]
+      "@recoil/*": ["src/recoil/*"],
+      "@lib/*": ["src/lib/*"]
     }
   },
   "include": ["next-env.d.ts", "**/*.ts", "**/*.tsx", "global.d.ts"],
